fix: errores ids invalidos

+++ src/main/java/acme/features/levelb/manager/project/ManagerProjectDeleteService.java
@@ -24,17 +24,23 @@ public class ManagerProjectDeleteService extends AbstractService<Manager, Projec
 
 	@Override
 	public void load() {
-		int id;
+		Integer id;
 
-		id = super.getRequest().getData("id", int.class);
-		this.project = this.repository.findProjectByIdWithDetails(id);
+		id = super.getRequest().getData("id", Integer.class, null);
+		if (id == null || id.intValue() == 0)
+			this.project = super.newObject(Project.class);
+		else {
+			this.project = this.repository.findProjectByIdWithDetails(id.intValue());
+			if (this.project == null)
+				this.project = super.newObject(Project.class);
+		}
 	}
 
 	@Override
 	public void authorise() {
 		boolean status;
 
-		status = this.project != null && this.project.getManager() != null && this.project.getManager().isPrincipal() && Boolean.TRUE.equals(this.project.getDraftMode());
+		status = this.project != null && this.project.getId() != 0 && this.project.getManager() != null && this.project.getManager().isPrincipal() && Boolean.TRUE.equals(this.project.getDraftMode());
 		super.setAuthorised(status);
 	}
 
@@ -47,6 +53,9 @@ public class ManagerProjectDeleteService extends AbstractService<Manager, Projec
 	public void validate() {
 		Long components;
 
+		if (this.project.getId() == 0)
+			return;
+
 		components = this.repository.countInventionsByProjectId(this.project.getId()) + this.repository.countCampaignsByProjectId(this.project.getId()) + this.repository.countStrategiesByProjectId(this.project.getId());
 		super.state(components == 0L, "draftMode", "acme.validation.project.has-components.message");
 	}
@@ -54,6 +63,9 @@ public class ManagerProjectDeleteService extends AbstractService<Manager, Projec
 	@Override
 	public void execute() {
 		Collection<ProjectMember> members;
+
+		if (this.project.getId() == 0)
+			return;
 
 		members = this.repository.findProjectMembersByProjectId(this.project.getId());
 		this.repository.deleteAll(members);
@@ -65,6 +77,7 @@ public class ManagerProjectDeleteService extends AbstractService<Manager, Projec
 		Tuple tuple;
 
 		tuple = super.unbindObject(this.project, "title", "keywords", "description", "kickOff", "closeOut", "draftMode");
+		tuple.put("id", this.project.getId() != 0 ? this.project.getId() : 0);
 		ProjectSupport.putDetails(tuple, this.project, this.repository);
 	}
